new permissions for the deployment user + ignore container updates (set via CI/CD)

+++ deploy/deploy_prod.sh
@@ -6,34 +6,38 @@ set -e
 # Navigate to the script's directory
 cd "$(dirname "$0")"
 
-DEV_TASK_DEF="course-management-dev"
-PROD_TASK_DEF="course-management-prod"
 
-echo "Deploying ${DEV_TASK_DEF} to prod environment"
+DEV_TAG=$1
 
-FILE_IN="${DEV_TASK_DEF}-current.json"
+if [ -z "$DEV_TAG" ]; then
+    echo "No tag provided. Fetching tag from the dev environment."
+    DEV_TASK_DEF="course-management-dev"
 
-echo "writing task definition to ${FILE_IN}"
-aws ecs describe-task-definition \
-    --task-definition ${DEV_TASK_DEF} \
-    > ${FILE_IN}
+    FILE_IN="${DEV_TASK_DEF}-current.json"
 
-DEV_TAG=$(
-    jq '.taskDefinition.containerDefinitions[].environment[] | select(.name == "VERSION").value' -r ${FILE_IN}
-)
+    echo "writing task definition to ${FILE_IN}"
+    aws ecs describe-task-definition \
+        --task-definition ${DEV_TASK_DEF} \
+        > ${FILE_IN}
+        
+    DEV_TAG=$(
+        jq '.taskDefinition.containerDefinitions[].environment[] | select(.name == "VERSION").value' -r ${FILE_IN}
+    )
+
+    rm ${FILE_IN}
+fi
+
 
 echo "deploying ${DEV_TAG} to prod"
 
-read -p "Are you sure you want to deploy to production? (y/N) " -n 1 -r
-echo
-
-# Check if the user's response is 'y' or 'Y'
-if [[ $REPLY =~ ^[Yy]$ ]]; then
-    echo "Okay, deploying..."
-else
-    # The user did not confirm, exit the script
-    echo "Deployment cancelled."
-    exit 1
+# Check if running outside GitHub Actions and need confirmation
+if [ -z "${GITHUB_ACTIONS}" ] && [ "${CONFIRM_DEPLOY}" != "true" ]; then
+    read -p "Are you sure you want to deploy to production? (y/N) " -n 1 -r
+    echo
+    if [[ ! $REPLY =~ ^[Yy]$ ]]; then
+        echo "Deployment cancelled."
+        exit 1
+    fi
 fi
 
 rm ${FILE_IN}
